+++ src/app/components/city/city.component.ts
@@ -108,6 +108,8 @@ handleDelete(id: any) {
     },
     error: (err: Error) => this.notifyService.showDanger(err.message, "Notification"),
   };
+  if(confirm("Are you sure to delete ")) {
   this.hotelService.deleteCity(id).subscribe(observer);
+  }
 }
 }
